#   refactor    : 파일 위치 변경 및 중복 파일 삭제
