+++ src/main/java/com/estsoft/project3/controller/LoginController.java
@@ -1,12 +1,13 @@
-package com.estsoft.project3.login.controller;
+package com.estsoft.project3.controller;
 
-import com.estsoft.project3.user.domain.User;
-import com.estsoft.project3.user.dto.SessionUser;
-import com.estsoft.project3.user.repository.UserRepository;
+import com.estsoft.project3.domain.User;
+import com.estsoft.project3.dto.SessionUser;
+import com.estsoft.project3.repository.UserRepository;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,14 +16,21 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class LoginController {
 
     private final UserRepository userRepository;
+    private final HttpSession httpSession;
 
-    public LoginController(UserRepository userRepository) {
+    public LoginController(UserRepository userRepository, HttpSession httpSession) {
         this.userRepository = userRepository;
+        this.httpSession = httpSession;
+    }
+
+    @GetMapping("/login")
+    public String showLoginPage() {
+        return "/login";
     }
 
     @GetMapping("/set-nickname")
     public String setNickname() {
-        return "/login/nickname";
+        return "/nickname";
     }
 
     @PostMapping("/set-nickname")
@@ -39,9 +47,14 @@ public class LoginController {
         return "redirect:/home";
     }
 
-    @GetMapping("/login")
-    public String showLoginPage() {
-        return "/login/login";
+    //추후 매핑 주소 변경해야함
+    @GetMapping("/main")
+    public String main(Model model) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        model.addAttribute("nickname", sessionUser.getNickname());
+        model.addAttribute("role", sessionUser.getRole());
+        model.addAttribute("score", sessionUser.getScore());
+        return "main";
     }
 
 }
