+++ src/main/java/com/estsoft/project3/repository/UserRepository.java
@@ -1,7 +1,9 @@
 package com.estsoft.project3.repository;
 
 import com.estsoft.project3.domain.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
 }
