+++ src/main/java/com/estsoft/project3/service/AllenApiService.java
@@ -11,8 +11,6 @@ import org.commonmark.parser.Parser;
 import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.node.Node;
 
-import java.util.Map;
-
 @Service
 public class AllenApiService {
