--- src/main/java/com/estsoft/project3/user/domain/Role.java
@@ -1,5 +1,0 @@
-package com.estsoft.project3.user.domain;
-
-public enum Role {
-    ROLE_ADMIN,ROLE_USER
-}
