--- src/main/java/com/estsoft/project3/user/domain/User.java
@@ -1,70 +1,0 @@
-package com.estsoft.project3.user.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@Table(name = "users")
-@Getter
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private long userId;
-
-    @Column(nullable = false)
-    private String provider;
-
-    @Column(nullable = false)
-    private String email;
-
-    private String nickname;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role;
-
-    @Column(nullable = false)
-    private String isActive;
-
-    private Long grade = 0L;
-    private Long score = 0L;
-
-    @CreatedDate
-    private LocalDateTime createdDate;
-
-    @LastModifiedDate
-    private LocalDateTime updatedDate;
-
-    @Builder
-    public User(String provider, String email, String nickname, Role role, String isActive) {
-        this.provider = provider;
-        this.email = email;
-        this.nickname = nickname;
-        this.role = role;
-        this.isActive = isActive;
-        this.grade = 0L;
-        this.score = 0L;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-}
