--- src/main/java/com/estsoft/project3/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.estsoft.project3.user.repository;
-
-import com.estsoft.project3.user.domain.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-
-}
