+++ src/main/resources/static/css/login.css
@@ -13,7 +13,7 @@ body {
 
 .login-card {
     background-color: white;
-    padding: 40px;
+    padding: 60px 80px;
     border-radius: 20px;
     box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
     text-align: center;
@@ -35,45 +35,4 @@ body {
 
 .google-btn img {
     height: 20px;
-}
-
-#nav-bar {
-    height: 64px;
-    border-bottom: 1px solid #ddd;
-    padding: 0 48px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-#nav-bar .d-flex {
-    height: 100%;
-}
-
-.logo {
-    height: 36px;
-    width: auto;
-    object-fit: contain;
-}
-
-.menu-link {
-    font-size: 16px;
-    color: #333;
-    text-decoration: none;
-    margin-left: 32px;
-    line-height: 64px; /* 수직 정렬 맞춤 */
-}
-
-.menu-link:hover {
-    color: #007bff;
-}
-
-.navbar-nickname {
-    font-weight: 500;
-    max-width: 150px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    line-height: 64px;
-    margin-left: 32px;
 }
