--- src/main/resources/templates/login/login.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>로그인</title>
-    <link th:href="@{/css/login/login.css}" rel="stylesheet">
-</head>
-<body>
-<div th:fragment="siteHeader">
-    <div id="nav-bar" class="bg-white border-bottom py-2 w-100">
-        <div class="d-flex justify-content-between align-items-center w-100 h-100 px-4">
-
-            <!-- Left Menu -->
-            <div class="d-flex align-items-center">
-                <a href="/home">
-                    <img src="/images/picture_logo.png" class="logo" alt="Logo"/>
-                </a>
-                <a class="menu-link ml-3" th:href="@{/review}">리뷰</a>
-            </div>
-
-            <!-- Right Menu -->
-            <div class="d-flex align-items-center">
-                <div th:if="${userId != null}">
-                    <span class="ml-3 d-inline-block text-truncate" th:text="${nickname}"
-                          th:title="${nickname}"></span>
-                    <a class="menu-link ml-3" th:href="@{/logout}">로그아웃</a>
-                </div>
-                <div th:if="${userId == null}">
-                    <a class="menu-link ml-3" th:href="@{/templates/login/login.html}">로그인</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<main class="login-container">
-    <div class="login-card">
-        <h2>Login</h2>
-        <a href="/oauth2/authorization/google" class="google-btn">
-            <img src="https://www.gstatic.com/firebasejs/ui/2.0.0/images/auth/google.svg"
-                 alt="Google logo"> Continue With Google
-        </a>
-    </div>
-</main>
-
-
-</body>
-</html>
