+++ src/main/resources/templates/nickname.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>닉네임 설정</title>
-  <link th:href="@{/css/login/login.css}" rel="stylesheet">
+  <link th:href="@{/css/login.css}" rel="stylesheet">
 </head>
 <body>
 <div th:fragment="siteHeader">
